feat: camera switching via NixOS specialisations
Replace runtime camera selection (config.txt dtoverlay) with
build-time device tree overlays and NixOS specialisations.
All three camera variants (imx462, imx296, imx477) are pre-built;
switching is instant via `switch-to-configuration boot` + reboot.

- Remove runtimeCameraSelection option, always bake overlay into DT
- Change default camera from imx296 to imx462
- Add specialisations to SD image (imx296, imx477) and netboot (imx296, imx462)
- Add pifinder-switch-camera wrapper script and sudoers entry
- Preserve camera choice across upgrades via /var/lib/pifinder/camera-type
- Delete Debian sys_utils.py, rename sys_utils_nixos.py to sys_utils.py
- Simplify utils.py get_sys_utils() (no more /etc/NIXOS check)
- Delegate callbacks.py get_camera_type() to sys_utils
- Deploy script: use FDTDIR in extlinux.conf, remove dtoverlay from config.txt
