+++ nixos/hardware.nix
@@ -12,18 +12,8 @@ in {
   options.pifinder = {
     cameraType = lib.mkOption {
       type = lib.types.enum [ "imx296" "imx462" "imx477" ];
-      default = "imx296";
+      default = "imx462";
       description = "Camera sensor type for PiFinder";
-    };
-
-    runtimeCameraSelection = lib.mkOption {
-      type = lib.types.bool;
-      default = true;
-      description = ''
-        Enable runtime camera selection via /boot/camera.txt.
-        When true, camera overlay is loaded via config.txt include (SD card).
-        When false, camera overlay is baked into device tree (netboot).
-      '';
     };
   };
 
@@ -82,8 +72,7 @@ in {
         '';
       };
 
-      # Camera dtoverlay — only bake into device tree when NOT using runtime selection
-      # For SD cards (runtimeCameraSelection=true), overlay is loaded via config.txt
+      # Camera dtoverlay — baked into device tree at build time
       cameraOverlay = {
         name = "${cameraDriver}-camera";
         dtboFile = "${pkgs.raspberrypifw}/share/raspberrypi/boot/overlays/${cameraDriver}.dtbo";
@@ -92,19 +81,8 @@ in {
       spi0Overlay
       uart3Overlay
       pwmOverlay
-    ] ++ lib.optionals (!cfg.runtimeCameraSelection) [
       cameraOverlay
     ];
-
-    # Runtime camera selection: create /boot/camera.txt with default overlay
-    # User can edit this file and reboot to switch cameras
-    environment.etc."camera.txt.default" = lib.mkIf cfg.runtimeCameraSelection {
-      text = ''
-        # Camera overlay - edit and reboot to change camera
-        # Options: imx296, imx290 (for imx462), imx477
-        dtoverlay=${cameraDriver}
-      '';
-    };
 
     # udev rules for hardware access without root
     services.udev.extraRules = ''
